+++ src/mesa/tnl_dd/t_dd_dmatmp.h
@@ -1,9 +1,8 @@
-
 /*
  * Mesa 3-D graphics library
- * Version:  3.5
+ * Version:  6.1
  *
- * Copyright (C) 1999-2001  Brian Paul   All Rights Reserved.
+ * Copyright (C) 1999-2004  Brian Paul   All Rights Reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
@@ -1203,7 +1202,7 @@ static GLboolean TAG(validate_render)( GLcontext *ctx,
 	    ok = GL_TRUE;
 	 } else if (HAVE_TRI_STRIPS && (ctx->_TriangleCaps & DD_FLATSHADE)) {
 	    if (HAVE_ELTS) {
-	       ok = count < GET_SUBSEQUENT_VB_MAX_ELTS();
+	       ok = (GLint) count < GET_SUBSEQUENT_VB_MAX_ELTS();
 	    }
 	    else {
 	       ok = GL_FALSE;
@@ -1216,7 +1215,7 @@ static GLboolean TAG(validate_render)( GLcontext *ctx,
 	 if (HAVE_QUADS) {
 	    ok = GL_TRUE;
 	 } else if (HAVE_ELTS) {
-	    ok = count < GET_SUBSEQUENT_VB_MAX_ELTS();
+	    ok = (GLint) count < GET_SUBSEQUENT_VB_MAX_ELTS();
 	 }
 	 else {
 	    ok = GL_FALSE;
